+++ pkg/providers/cloudlets/data_akamai_cloudlets_api_prioritization_match_rule.go
@@ -192,7 +192,7 @@ func dataSourceCloudletsAPIPrioritizationMatchRule() *schema.Resource {
 func dataSourceCloudletsAPIPrioritizationMatchRuleRead(_ context.Context, d *schema.ResourceData, _ interface{}) diag.Diagnostics {
 	matchRulesList, err := tf.GetListValue("match_rules", d)
 	if err != nil {
-		return diag.FromErr(err)
+		return handleEmptyMatchRules(err, d, "data_akamai_cloudlets_api_prioritization_match_rule")
 	}
 
 	if err := setMatchRuleSchemaType(matchRulesList, cloudlets.MatchRuleTypeAP); err != nil {