+++ pkg/providers/cloudlets/data_akamai_cloudlets_audience_segmentation_match_rule.go
@@ -215,7 +215,7 @@ func dataSourceCloudletsAudienceSegmentationMatchRule() *schema.Resource {
 func dataSourceCloudletsAudienceSegmentationMatchRuleRead(_ context.Context, d *schema.ResourceData, _ interface{}) diag.Diagnostics {
 	matchRulesList, err := tf.GetListValue("match_rules", d)
 	if err != nil {
-		return diag.FromErr(err)
+		return handleEmptyMatchRules(err, d, "data_akamai_cloudlets_audience_segmentation_match_rule")
 	}
 
 	if err = setMatchRuleSchemaType(matchRulesList, cloudlets.MatchRuleTypeAS); err != nil {