+++ pkg/providers/cloudlets/data_akamai_cloudlets_audience_segmentation_match_rule_test.go
@@ -12,30 +12,42 @@ func TestDataCloudletsAudienceSegmentationMatchRule(t *testing.T) {
 	tests := map[string]struct {
 		configPath       string
 		expectedJSONPath string
+		matchRulesSize   int
+		emptyRules       bool
 	}{
 		"basic valid rule set": {
 			configPath:       "testdata/TestDataCloudletsAudienceSegmentationMatchRule/basic.tf",
 			expectedJSONPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/rules/basic_rules.json",
+			matchRulesSize:   1,
 		},
 		"valid rule set with simple value": {
 			configPath:       "testdata/TestDataCloudletsAudienceSegmentationMatchRule/omv_simple.tf",
 			expectedJSONPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/rules/omv_simple_rules.json",
+			matchRulesSize:   1,
 		},
 		"valid rule set with object value": {
 			configPath:       "testdata/TestDataCloudletsAudienceSegmentationMatchRule/omv_object.tf",
 			expectedJSONPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/rules/omv_object_rules.json",
+			matchRulesSize:   1,
 		},
 		"valid rule set with range value": {
 			configPath:       "testdata/TestDataCloudletsAudienceSegmentationMatchRule/omv_range.tf",
 			expectedJSONPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/rules/omv_range_rules.json",
+			matchRulesSize:   1,
 		},
 		"valid rule set without value": {
 			configPath:       "testdata/TestDataCloudletsAudienceSegmentationMatchRule/omv_empty.tf",
 			expectedJSONPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/rules/omv_empty_rules.json",
+			matchRulesSize:   1,
 		},
 		"valid complex rule set": {
 			configPath:       "testdata/TestDataCloudletsAudienceSegmentationMatchRule/omv_complex.tf",
 			expectedJSONPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/rules/omv_complex_rules.json",
+			matchRulesSize:   3,
+		},
+		"no match rules": {
+			configPath: "testdata/TestDataCloudletsAudienceSegmentationMatchRule/no_match_rules.tf",
+			emptyRules: true,
 		},
 	}
 
@@ -46,13 +58,8 @@ func TestDataCloudletsAudienceSegmentationMatchRule(t *testing.T) {
 				Steps: []resource.TestStep{
 					{
 						Config: testutils.LoadFixtureString(t, test.configPath),
-						Check: resource.ComposeAggregateTestCheckFunc(
-							resource.TestCheckResourceAttr(
-								"data.akamai_cloudlets_audience_segmentation_match_rule.test", "json",
-								testutils.LoadFixtureString(t, test.expectedJSONPath)),
-							resource.TestCheckResourceAttr(
-								"data.akamai_cloudlets_audience_segmentation_match_rule.test", "match_rules.0.type", "asMatchRule"),
-						),
+						Check: checkMatchRulesAttr(t, "asMatchRule", "data.akamai_cloudlets_audience_segmentation_match_rule.test",
+							test.expectedJSONPath, test.emptyRules, test.matchRulesSize),
 					},
 				},
 			})