+++ pkg/providers/cloudlets/data_akamai_cloudlets_phased_release_match_rule.go
@@ -213,7 +213,7 @@ func dataSourceCloudletsPhasedReleaseMatchRule() *schema.Resource {
 func dataSourcePhasedReleaseMatchRuleRead(_ context.Context, d *schema.ResourceData, _ interface{}) diag.Diagnostics {
 	matchRulesList, err := tf.GetListValue("match_rules", d)
 	if err != nil {
-		return diag.FromErr(err)
+		return handleEmptyMatchRules(err, d, "data_akamai_cloudlets_phased_release_match_rule")
 	}
 
 	if err = setMatchRuleSchemaType(matchRulesList, cloudlets.MatchRuleTypePR); err != nil {