+++ pkg/providers/cloudlets/data_akamai_cloudlets_request_control_match_rule.go
@@ -193,7 +193,7 @@ func dataSourceCloudletsRequestControlMatchRule() *schema.Resource {
 func dataSourceCloudletsRequestControlMatchRuleRead(_ context.Context, d *schema.ResourceData, _ interface{}) diag.Diagnostics {
 	matchRulesList, err := tf.GetListValue("match_rules", d)
 	if err != nil {
-		return diag.FromErr(err)
+		return handleEmptyMatchRules(err, d, "data_akamai_cloudlets_request_control_match_rule")
 	}
 
 	if err := setMatchRuleSchemaType(matchRulesList, cloudlets.MatchRuleTypeRC); err != nil {