+++ pkg/providers/cloudlets/data_akamai_cloudlets_visitor_prioritization_match_rule_test.go
@@ -17,6 +17,7 @@ func TestDataCloudletsVisitorPrioritizationMatchRule(t *testing.T) {
 		configPath       string
 		expectedJSONPath string
 		matchRulesSize   int
+		emptyRules       bool
 	}{
 		"valid all vars map": {
 			configPath:       fmt.Sprintf("%s/vars_map.tf", workdir),
@@ -38,6 +39,10 @@ func TestDataCloudletsVisitorPrioritizationMatchRule(t *testing.T) {
 			expectedJSONPath: fmt.Sprintf("%s/rules/omv_object_rules.json", workdir),
 			matchRulesSize:   1,
 		},
+		"no match rules": {
+			configPath: fmt.Sprintf("%s/no_match_rules.tf", workdir),
+			emptyRules: true,
+		},
 	}
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
@@ -46,15 +51,8 @@ func TestDataCloudletsVisitorPrioritizationMatchRule(t *testing.T) {
 				Steps: []resource.TestStep{
 					{
 						Config: testutils.LoadFixtureString(t, test.configPath),
-						Check: resource.ComposeAggregateTestCheckFunc(
-							resource.TestCheckResourceAttr(
-								"data.akamai_cloudlets_visitor_prioritization_match_rule.test", "json",
-								testutils.LoadFixtureString(t, test.expectedJSONPath)),
-							resource.TestCheckResourceAttr(
-								"data.akamai_cloudlets_visitor_prioritization_match_rule.test", "match_rules.0.type", "vpMatchRule"),
-							resource.TestCheckResourceAttr(
-								"data.akamai_cloudlets_visitor_prioritization_match_rule.test", "match_rules.#", strconv.Itoa(test.matchRulesSize)),
-						),
+						Check: checkMatchRulesAttr(t, "vpMatchRule", "data.akamai_cloudlets_visitor_prioritization_match_rule.test",
+							test.expectedJSONPath, test.emptyRules, test.matchRulesSize),
 					},
 				},
 			})