+++ pkg/providers/cloudlets/testdata/TestDataCloudletsAudienceSegmentationMatchRule/no_match_rules.tf
@@ -1,0 +1,5 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_cloudlets_audience_segmentation_match_rule" "test" {}