+++ pkg/providers/cloudlets/testdata/TestDataCloudletsEdgeRedirectorMatchRule/no_match_rules.tf
@@ -1,0 +1,5 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_cloudlets_edge_redirector_match_rule" "test" {}