+++ pkg/providers/cloudlets/testdata/TestDataCloudletsLoadBalancerMatchRule/no_match_rules.tf
@@ -1,0 +1,5 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_cloudlets_application_load_balancer_match_rule" "test" {}