+++ pkg/providers/cloudlets/testdata/TestDataCloudletsPhasedReleaseMatchRule/no_match_rules.tf
@@ -1,0 +1,5 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_cloudlets_phased_release_match_rule" "test" {}