+++ pkg/providers/cloudlets/testdata/TestDataCloudletsVisitorPrioritizationMatchRule/no_match_rules.tf
@@ -1,0 +1,5 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_cloudlets_visitor_prioritization_match_rule" "test" {}